add fe flow for duplicate name issue
